Update cron to be executed every 5 minutes

+++ www/cron.php
@@ -6,30 +6,35 @@
 require_once __DIR__.'/config.inc.php';
 require_once __DIR__.'/funcs.inc.php';
 
+$go = getopt('h', ['hourly']);
+$hourly = isset($go['h']) || isset($go['hourly']) || isset($_GET['h']) || isset($_GET['hourly']);
+
 # Set tasks in error when their nb_fails is too high
 $stmt = $db->prepare("UPDATE `queue` SET status='error' WHERE status='queued' AND nb_fails>=:maxfails;");
 $stmt->execute(array(':maxfails' => $CFG_max_fails));
 
-# Delete old tasks and logs
-$stmt = $db->prepare("DELETE FROM `queue` WHERE received_time <= NOW() - INTERVAL :days day AND grading_start_time <= NOW() - INTERVAL :days day;");
-$stmt->execute(array(':days' => $CFG_keep_old_days));
-$stmt = $db->prepare("DELETE FROM `done` WHERE received_time <= NOW() - INTERVAL :days day AND grading_start_time <= NOW() - INTERVAL :days day;");
-$stmt->execute(array(':days' => $CFG_keep_old_days));
-$stmt = $db->prepare("DELETE FROM `log` WHERE datetime <= NOW() - INTERVAL :days day;");
-$stmt->execute(array(':days' => $CFG_keep_old_days));
+if($hourly) {
+  # Delete old tasks and logs
+  $stmt = $db->prepare("DELETE FROM `queue` WHERE received_time <= NOW() - INTERVAL :days day AND grading_start_time <= NOW() - INTERVAL :days day;");
+  $stmt->execute(array(':days' => $CFG_keep_old_days));
+  $stmt = $db->prepare("DELETE FROM `done` WHERE received_time <= NOW() - INTERVAL :days day AND grading_start_time <= NOW() - INTERVAL :days day;");
+  $stmt->execute(array(':days' => $CFG_keep_old_days));
+  $stmt = $db->prepare("DELETE FROM `log` WHERE datetime <= NOW() - INTERVAL :days day;");
+  $stmt->execute(array(':days' => $CFG_keep_old_days));
 
-# Delete orphan job_types
-$db->query("DELETE FROM `job_types` WHERE NOT EXISTS (SELECT 1 FROM `queue` WHERE `job_types`.`jobid`=`queue`.`id`);");
+  # Delete orphan job_types
+  $db->query("DELETE FROM `job_types` WHERE NOT EXISTS (SELECT 1 FROM `queue` WHERE `job_types`.`jobid`=`queue`.`id`);");
+}
 
 # Warn about tasks in error / stuck
 $stmt = $db->prepare("
     SELECT COUNT(*) FROM `queue`
-    WHERE received_time >= NOW() - INTERVAL :hours hour - INTERVAL :cronintv hour
-    AND received_time <= NOW() - INTERVAL :hours hour
+    WHERE received_time >= NOW() - INTERVAL 5 minute
+    AND received_time <= NOW() - INTERVAL :seconds second
     AND (status = 'queued' OR status = 'error');");
-$stmt->execute(array(':hours' => $CFG_warn_hours, ':cronintv' => $CFG_cron_interval_hours));
+$stmt->execute(array(':seconds' => $CFG_warn_seconds));
 $row = $stmt->fetch();
-if($row[0] > 0) {
+if($row[0] > $CFG_warn_nb) {
   $msg = strtr("
 Hi,
 
@@ -52,14 +57,15 @@ graderqueue", array(':count' => $row[0], ':url' => $CFG_interface_url));
 # Try waking-up servers again
 wake_up_server_by_error();
 
-# Warn about servers not waking-up
-$stmt = $db->query("SELECT name, wakeup_fails, last_poll_time FROM `servers` WHERE wakeup_fails >= 3;");
-$wakeup_error_servers = '';
-while($row = $stmt->fetch()) {
-  $wakeup_error_servers .= $row['name'] . "(" . $row['wakeup_fails'] . " failures, last poll " . $row['last_poll_time'] . ")\n";
-}
-if($wakeup_error_servers != '') {
-  $msg = strtr("
+if($hourly) {
+  # Warn about servers not waking-up
+  $stmt = $db->query("SELECT name, wakeup_fails, last_poll_time FROM `servers` WHERE wakeup_fails >= 3;");
+  $wakeup_error_servers = '';
+  while($row = $stmt->fetch()) {
+    $wakeup_error_servers .= $row['name'] . "(" . $row['wakeup_fails'] . " failures, last poll " . $row['last_poll_time'] . ")\n";
+  }
+  if($wakeup_error_servers != '') {
+    $msg = strtr("
 Hi,
 
 The following servers haven't woken up after a few tries:
@@ -70,12 +76,13 @@ Cheers,
 
 --
 graderqueue", array(':errorserv' => $wakeup_error_servers, ':url' => $CFG_interface_url));
-  if(gettype($CFG_admin_email) == 'array') {
-    foreach($CFG_admin_email as $recipient) {
-      mail($recipient, "[graderqueue] Servers not waking-up", $msg);
+    if(gettype($CFG_admin_email) == 'array') {
+      foreach($CFG_admin_email as $recipient) {
+        mail($recipient, "[graderqueue] Servers not waking-up", $msg);
+      }
+    } else {
+      mail($CFG_admin_email, "[graderqueue] Servers not waking-up", $msg);
     }
-  } else {
-    mail($CFG_admin_email, "[graderqueue] Servers not waking-up", $msg);
   }
 }
 ?>
